Fix a bug in `.toHz()` which caused incorrect orders of magnitude.
Summary:
If you gave a frequency, "14.010.000", `.toHz()` would take each of those
segments and convert them to a Long, after doing some checks on them.

Converting `"010"` to a Long effectively returns `10`.

If you plug that back into the frequency, you'd get `"14.100.000"` which is
obviously not correct.

We fix this by storing the originally split `List[String]` and a separate one
which maps them to `Long`.

Test Plan: Added a testcase to cover this and ran the suite successfully.

Reviewers: jcarter

Reviewed By: jcarter

Differential Revision: https://phabricator.edgesysdesign.com/D29