Disable numeric sanity check for now.
Summary:
We currently throw when you try to do `1.MHz + (- 20.kHz)`, but yet 1.MHz` and
`-20.kHz` are both valid, and both able to be converted to Hertz.

`-20.kHz` alone also worked, but doing addition/subtraction/etc required a
conversion to hertz, which is where we threw.

The reason we throw is because we sanity check that frequencies can only contain
`.` and numbers, but this is obviously a wrong assumption.

The correct fix is to do the sanity check, but allow the first character to be
either - or a number, but this is somewhat more complex and would probably
require a regex to do sanely, which is slightly expensive.

This diff removes the sanity check altogether. The tradeoff is that while now we
can evaluate `1.MHz + (-20.kHz)`, if you type `"100.f123".MHz`, we end up
passing that to `BigInt`, which throws an ugly `NumberFormatException` with a
huge stack trace. I think this is a fair tradeoff, however because the error
still roughly makes sense and if you're putting letters in your frequency, it
probably means you're not validating user input properly on the client side
anyway. I think this is better than making assumptions on behalf of the
developer.

Test Plan:
  scala> "10.f30.132".MHz
  java.lang.NumberFormatException: For input string: "f30"
  	at java.lang.NumberFormatException.forInputString(NumberFormatException.java:65)
  	at java.lang.Integer.parseInt(Integer.java:492)

  scala> 0.MHz + (- 1.MHz)
    res3: com.edgesysdesign.frequency.Frequency = -1.000.000

Reviewers: jcarter

Reviewed By: jcarter

Maniphest Tasks: T85

Differential Revision: https://phabricator.edgesysdesign.com/D84